move more sample components and some sample only utils to /samples/utils.ts
