+++ packages/components/src/index_samples.ts
@@ -1,5 +1,7 @@
 import { SampleTypeBasePage } from './internal/samples/SampleTypeBasePage';
 
+import { getSampleWizardURL, filterSampleRowsForOperation, getSampleSetMenuItem, isFindByIdsSchema } from './internal/samples/utils';
+
 // TODO should these files be moved?
 import { SampleActionsButton } from './internal/components/samples/SampleActionsButton';
 import { SampleAliquotsGridPanel } from './internal/components/samples/SampleAliquotsGridPanel';
@@ -12,18 +14,36 @@ import { SampleSetDeleteModal } from './internal/components/samples/SampleSetDel
 import { SampleSetSummary } from './internal/components/samples/SampleSetSummary';
 import { SamplesDeriveButtonBase } from './internal/components/samples/SamplesDeriveButtonBase';
 import { SamplesEditButton } from './internal/components/samples/SamplesEditButton';
+import { SampleAliquotDetailHeader } from './internal/components/samples/SampleAliquotDetailHeader';
+import { SampleCreationTypeModal } from './internal/components/samples/SampleCreationTypeModal';
+import { CreateSamplesSubMenu } from './internal/components/samples/CreateSamplesSubMenu';
+import { CreateSamplesSubMenuBase } from './internal/components/samples/CreateSamplesSubMenuBase';
+import { SamplesAssayButton } from './internal/components/samples/SamplesAssayButton';
+import { SampleTypeInsightsPanel } from './internal/components/samples/SampleTypeInsightsPanel';
+import { EntityCrossProjectSelectionConfirmModal } from './internal/components/entities/EntityCrossProjectSelectionConfirmModal';
 
 export {
+    filterSampleRowsForOperation,
+    getSampleSetMenuItem,
+    getSampleWizardURL,
+    isFindByIdsSchema,
+    CreateSamplesSubMenu,
+    CreateSamplesSubMenuBase,
+    EntityCrossProjectSelectionConfirmModal,
     SampleActionsButton,
+    SampleAliquotDetailHeader,
     SampleAliquotsGridPanel,
     SampleAliquotsSummary,
     SamplesAddButton,
     SampleAssayDetail,
+    SampleCreationTypeModal,
     SampleDetailEditing,
     SampleLineageGraph,
     SampleSetDeleteModal,
     SampleSetSummary,
     SampleTypeBasePage,
+    SampleTypeInsightsPanel,
+    SamplesAssayButton,
     SamplesDeriveButtonBase,
     SamplesEditButton,
 };
